:recycle: Refatoração de código do votify-console
