+++ votify-console/src/main/java/br/com/votify/console/menus/MainMenu.java
@@ -3,9 +3,9 @@ package br.com.votify.console.menus;
 public class MainMenu extends OptionsMenu {
     public MainMenu() {
         menuCollection.addMenu(this);
+        menuCollection.addMenu(new AuthMenu());
         menuCollection.addMenu(new UsersMenu());
         menuCollection.addMenu(new PollsMenu());
-        menuCollection.addMenu(new UserContextMenu());
     }
 
     @Override
