+++ votify-console/src/main/java/br/com/votify/console/menus/OptionsMenu.java
@@ -14,7 +14,6 @@ public abstract class OptionsMenu extends Menu {
         List<Menu> menus = menuCollection.getMenus();
         int option = 0;
         while (option != menus.size()) {
-            ConsoleUtils.clear();
             printBanner();
 
             for (int i = 1; i < menus.size(); i++) {
@@ -34,29 +33,6 @@ public abstract class OptionsMenu extends Menu {
                 }
             }
         }
-    }
-
-    private void printBanner() {
-        String optionName = getOptionName();
-        StringBuilder stringBuilder = new StringBuilder();
-        for (int i = 0; i < Math.max(46, optionName.length()); i++) {
-            if (i % 2 == 0) {
-                stringBuilder.append('-');
-                continue;
-            }
-            stringBuilder.append('=');
-        }
-        String bannerDecorator = stringBuilder.toString();
-        stringBuilder.delete(0, bannerDecorator.length());
-
-        int spaceOnEachEdge = (bannerDecorator.length() - optionName.length()) / 2;
-        stringBuilder.append(" ".repeat(Math.max(0, spaceOnEachEdge)));
-        stringBuilder.append(optionName);
-        stringBuilder.append(" ".repeat(Math.max(0, spaceOnEachEdge)));
-
-        System.out.println(bannerDecorator);
-        System.out.println(stringBuilder);
-        System.out.println(bannerDecorator);
     }
 
     protected abstract String getExitName();
