+++ votify-console/src/main/java/br/com/votify/console/menus/UsersMenu.java
@@ -1,17 +1,15 @@
 package br.com.votify.console.menus;
 
+import br.com.votify.console.menus.users.CurrentUserMenu;
+import br.com.votify.console.menus.users.DeleteUserContextMenu;
 import br.com.votify.console.menus.users.GetByIdUserMenu;
-import br.com.votify.console.menus.users.LoginUserMenu;
-import br.com.votify.console.menus.users.RegisterUserMenu;
-import br.com.votify.console.menus.users.ResetPasswordUserMenu;
 
 public class UsersMenu extends OptionsMenu {
     public UsersMenu() {
         menuCollection.addMenu(this);
-        menuCollection.addMenu(new RegisterUserMenu());
-        menuCollection.addMenu(new LoginUserMenu());
-        menuCollection.addMenu(new ResetPasswordUserMenu());
         menuCollection.addMenu(new GetByIdUserMenu());
+        menuCollection.addMenu(new CurrentUserMenu());
+        menuCollection.addMenu(new DeleteUserContextMenu());
     }
 
     @Override
