+++ votify-console/src/main/java/br/com/votify/console/menus/auth/LoginMenu.java
@@ -1,26 +1,32 @@
-package br.com.votify.console.menus.context;
+package br.com.votify.console.menus.auth;
 
 import br.com.votify.console.callers.VotifyApiCaller;
 import br.com.votify.console.menus.Menu;
 import br.com.votify.console.utils.ConsoleUtils;
 import br.com.votify.dto.ApiResponse;
-import br.com.votify.dto.users.UserDetailedViewDTO;
+import br.com.votify.dto.users.UserLoginDTO;
 
 import java.util.Scanner;
 
-public class CurrentUserContextMenu extends Menu {
+public class LoginMenu extends Menu {
     private final Scanner scanner = new Scanner(System.in);
 
     @Override
     public void run() {
-        ConsoleUtils.clear();
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-        System.out.println("              Minhas informações               ");
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+        printBanner();
 
-        ApiResponse<UserDetailedViewDTO> response = VotifyApiCaller.CONTEXT.getUser();
+        UserLoginDTO dto = new UserLoginDTO();
+
+        System.out.print("Insira o e-mail: ");
+        dto.setEmail(scanner.nextLine());
+
+        System.out.print("Insira a senha: ");
+        dto.setPassword(scanner.nextLine());
+
+        ApiResponse<?> response = VotifyApiCaller.AUTH.login(dto);
+        ConsoleUtils.clear();
         if (response.isSuccess()) {
-            System.out.println("Suas informações: " + response.getData());
+            System.out.println("Login feito com sucesso");
         } else {
             System.out.println("Erro: " + response.getErrorMessage());
         }
@@ -29,6 +35,6 @@ public class CurrentUserContextMenu extends Menu {
 
     @Override
     public String getOptionName() {
-        return "Minhas informações";
+        return "Fazer login";
     }
 }
