+++ votify-console/src/main/java/br/com/votify/console/menus/auth/LogoutMenu.java
@@ -1,4 +1,4 @@
-package br.com.votify.console.menus.context;
+package br.com.votify.console.menus.auth;
 
 import br.com.votify.console.callers.VotifyApiCaller;
 import br.com.votify.console.menus.Menu;
@@ -7,20 +7,17 @@ import br.com.votify.dto.ApiResponse;
 
 import java.util.Scanner;
 
-public class LogoutUserContextMenu extends Menu {
+public class LogoutMenu extends Menu {
     private final Scanner scanner = new Scanner(System.in);
 
     @Override
     public void run() {
-        ConsoleUtils.clear();
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-        System.out.println("                     Logout                    ");
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+        printBanner();
         System.out.print("Tem certeza? [S ou N]: ");
 
         if (!ConsoleUtils.getBooleanFromInput(scanner)) return;
 
-        ApiResponse<?> response = VotifyApiCaller.USERS.logout();
+        ApiResponse<?> response = VotifyApiCaller.AUTH.logout();
         if (response.isSuccess()) {
             System.out.println("Deslogado com sucesso");
         } else {
