+++ votify-console/src/main/java/br/com/votify/console/menus/auth/RefreshTokensMenu.java
@@ -1,24 +1,20 @@
-package br.com.votify.console.menus.context;
+package br.com.votify.console.menus.auth;
 
 import br.com.votify.console.callers.VotifyApiCaller;
 import br.com.votify.console.menus.Menu;
 import br.com.votify.console.utils.ConsoleUtils;
 import br.com.votify.dto.ApiResponse;
-import br.com.votify.dto.users.UserDetailedViewDTO;
 
 import java.util.Scanner;
 
-public class RegenerateTokensUserContextMenu extends Menu {
+public class RefreshTokensMenu extends Menu {
     private final Scanner scanner = new Scanner(System.in);
 
     @Override
     public void run() {
-        ConsoleUtils.clear();
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-        System.out.println("               Regenerar tokens                ");
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+        printBanner();
 
-        ApiResponse<?> response = VotifyApiCaller.CONTEXT.regenerateTokens();
+        ApiResponse<?> response = VotifyApiCaller.AUTH.refreshTokens();
         if (response.isSuccess()) {
             System.out.println("Tokens regenerados com sucesso");
         } else {
