+++ votify-console/src/main/java/br/com/votify/console/menus/auth/RegisterMenu.java
@@ -1,35 +1,39 @@
-package br.com.votify.console.menus.users;
+package br.com.votify.console.menus.auth;
 
 import br.com.votify.console.callers.VotifyApiCaller;
 import br.com.votify.console.menus.Menu;
 import br.com.votify.console.utils.ConsoleUtils;
 import br.com.votify.dto.ApiResponse;
-import br.com.votify.dto.users.UserLoginDTO;
+import br.com.votify.dto.users.UserDetailedViewDTO;
+import br.com.votify.dto.users.UserRegisterDTO;
 
 import java.util.Scanner;
 
-public class LoginUserMenu extends Menu {
+public class RegisterMenu extends Menu {
     private final Scanner scanner = new Scanner(System.in);
 
     @Override
     public void run() {
-        ConsoleUtils.clear();
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-        System.out.println("                  Fazer login                  ");
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+        printBanner();
 
-        UserLoginDTO dto = new UserLoginDTO();
+        UserRegisterDTO dto = new UserRegisterDTO();
 
         System.out.print("Insira o e-mail: ");
         dto.setEmail(scanner.nextLine());
 
+        System.out.print("Insira o nome: ");
+        dto.setName(scanner.nextLine());
+
+        System.out.print("Insira o nome de usuário: ");
+        dto.setUserName(scanner.nextLine());
+
         System.out.print("Insira a senha: ");
         dto.setPassword(scanner.nextLine());
 
-        ApiResponse<?> response = VotifyApiCaller.USERS.login(dto);
+        ApiResponse<UserDetailedViewDTO> response = VotifyApiCaller.AUTH.register(dto);
         ConsoleUtils.clear();
         if (response.isSuccess()) {
-            System.out.println("Login feito com sucesso");
+            System.out.println("Usuário criado com sucesso");
         } else {
             System.out.println("Erro: " + response.getErrorMessage());
         }
@@ -38,6 +42,6 @@ public class LoginUserMenu extends Menu {
 
     @Override
     public String getOptionName() {
-        return "Fazer login";
+        return "Registrar usuário";
     }
 }
