+++ votify-console/src/main/java/br/com/votify/console/menus/auth/ResetPasswordMenu.java
@@ -1,4 +1,4 @@
-package br.com.votify.console.menus.users;
+package br.com.votify.console.menus.auth;
 
 import br.com.votify.console.callers.VotifyApiCaller;
 import br.com.votify.console.menus.Menu;
@@ -10,22 +10,19 @@ import br.com.votify.dto.users.PasswordResetResponseDTO;
 
 import java.util.Scanner;
 
-public class ResetPasswordUserMenu extends Menu {
+public class ResetPasswordMenu extends Menu {
     private final Scanner scanner = new Scanner(System.in);
 
     @Override
     public void run() {
-        ConsoleUtils.clear();
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-        System.out.println("                Esqueci a senha                ");
-        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+        printBanner();
 
         PasswordResetRequestDTO dto1 = new PasswordResetRequestDTO();
 
         System.out.print("Insira o e-mail: ");
         dto1.setEmail(scanner.nextLine());
 
-        ApiResponse<PasswordResetResponseDTO> response1 = VotifyApiCaller.PASSWORD.forgot(dto1);
+        ApiResponse<PasswordResetResponseDTO> response1 = VotifyApiCaller.AUTH.forgotPassword(dto1);
         if (!response1.isSuccess()) {
             System.out.println("Erro: " + response1.getErrorMessage());
             ConsoleUtils.pressEnterToContinue(scanner);
@@ -41,7 +38,7 @@ public class ResetPasswordUserMenu extends Menu {
         System.out.print("Insira a nova senha: ");
         dto2.setNewPassword(scanner.nextLine());
 
-        ApiResponse<?> response2 = VotifyApiCaller.PASSWORD.reset(dto2);
+        ApiResponse<?> response2 = VotifyApiCaller.AUTH.resetPassword(dto2);
         ConsoleUtils.clear();
         if (response2.isSuccess()) {
             System.out.println("Reset de senha feito com sucesso");
